support custom faraday
Pass thru and use a faraday object specified be the embedding.
This enables us to use the ':rake' faraday adpater in helltool tests
so that we can make calls to core via the sdk in tests that flow
through rake-test and not real http.

Also, ensure that swagger loading doesn't do the wrong thing when
there is a 404 with a body.