Work in progress of streaming support.
This adds support to pass a block to sdk calls. When the block is
present the sdk will stream the raw body from the server to the block.

** This is experimental **

The intended use case is when receiving huge query results from the server
that may not all fit in memory and need to be streamed to disk.

The current Faraday lacks streaming support. So, this code uses Net::HTTP
rather than Faraday *if* a block is passed for a call to the sdk.
There are plenty of cases where this may not work well (or not work at all).
But, it could be useful in many cases. Time will tell.

This work is not yet finished and is almost completely untested.
Improvements are on the way.