Don't hand complete URL to Net::HTTP::Foo classes
Was failing to remove the leading scheme:host:port
This had awesomely odd behavior with with Rack/Sinatra parsing on
the server. It would parse out the correct path and route the request.
But, it failed to find the query part of the URL. So, endpoints were
not getting the query part at all.