Use Makefile for multiple ruby version testing
rbenv is the newer standard for Looker ruby version management. shelling
out from a rake command to "bundle exec" was not executing the tests
with the correct ruby version due to the
[environment changes it makes](https://bundler.io/man/bundle-exec.1.html#ENVIRONMENT-MODIFICATIONS)

Note: I could not get jruby-1.7.19 to build on the mac so I did not
include it.

added Google LLC to copyright notice on Makefile